requirements.txt
einops==0.8.0
fitz==0.0.1.dev2
flasgger==0.9.7.1
Flask==3.1.0
Flask_Caching==2.3.0
Flask_Cors==5.0.0
Flask_HTTPAuth==4.8.0
Flask_RESTful==0.3.10
matplotlib==3.10.0
numpy==2.2.1
onnx==1.17.0
onnxruntime==1.19.2
onnxsim==0.4.36
opencv_python==4.10.0.84
Pillow==11.1.0
pycocotools==2.0.8
PyMySQL==1.1.1
scipy==1.14.1
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.0
ultralytics==8.3.29
Werkzeug==3.1.3
